refactor: js -> jshelper
