+++ plugin/internal/jshelper/module_test.go
@@ -1,4 +1,4 @@
-package js
+package jshelper
 
 import (
 	"testing"
