+++ plugin/internal/jshelper/transpilation.go
@@ -1,4 +1,4 @@
-package js
+package jshelper
 
 import (
 	"errors"
