+++ plugin/internal/jshelper/transpilation_test.go
@@ -1,4 +1,4 @@
-package js
+package jshelper
 
 import (
 	"testing"
